requirements.txt
# langchain core
langchain
langchain-core

# openai integration
langchain-openai
openai

# anthropic integration
langchain-anthropic

# google gemini (PaLM) integration
langchain-google-genai
google-generativeai

# hugging face integration
langchain-huggingface
transformers
huggingface-hub

# environment variable management
python-dotenv

# machine learning utilities
numpy
scikit-learn

